+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - FIX: AccountRenewalCountMax and DomainRenewalCountMax bumped at configuration update
 - FIX: fix cli tests
+- FIX: fix domain renewal
 
 ## v0.4.4